Extend PATH with directory of executable
This is actually THE fix for cx_Freeze 5. The problem is that when
modules are packaged outside ZIP archive (which is the default), _gi.pyd
is placed in the gi subdirectory. _gi.pyd depends on some DLLs, so an
attempt to launch the app outside the directory of DLLs (that is, the
directory of the main executable) will give a "DLL load failed" error.

To fix this, we should extend the PATH manually since cx_Freeze doesn't
do this.