export: Better string formatting
Fixes various issues of the naive implementation. In particular, it's now
possible to insert a single % (using %%) before a character that's a format
specifier without triggering the substitution.

The tooltiop string for % specifier is also added to gettext. Although there
is nothing to translate, it may be required to change the order for RTL text.